Rename misspelled package
