+++ query/src/test/java/org/infinispan/query/partition/NonSharedIndexTest.java
@@ -1,4 +1,4 @@
-package org.infinispan.query.paritition;
+package org.infinispan.query.partition;
 
 import java.util.List;
